ORCA/src/otdd/requirements.txt

pandas
scikit-learn

numpy
scipy
matplotlib
torchvision
torchtext
sentence-transformers
skorch

pot

cmake
pykeops
git+https://github.com/jeanfeydy/geomloss
munkres

tqdm
addict
seaborn
adjustText
h5py
opentsne

watermark
seaborn
celluloid
git+https://github.com/rossant/ipycache
